Introduce SafeEventEmitterProvider interface
In some controller tests it's useful to create a fake provider that
conforms to the same interface as a "real" provider but doesn't actually
hit the network. To make such a fake provider, however, we must create a
class that extends SafeEventEmitterProvider. This is unfortunate as it
introduces the possibility that the network could be hit accidentally if
not all of the methods in the interface are appropriately overridden.

This commit introduces an interface that allows us to _implement_ the
same interface without having to extend it.